cgi: work around a bug in NetSurf's <BASE HREF> support
It must have a hostname otherwise it will be ignored. We will at least
assume HTTP_HOST is present instead of trying to fall back on
SERVER_NAME and SERVER_PORT for simplicity sake. NetSurf also has the
peculiar behavior that a protocol-relative BASE HREF URI is rewritten to
be HTTP even if the document URI is HTTPS, but whatever.